--- src/main/java/roomescape/reservation/controller/ReservationTimeController.java
@@ -1,52 +1,0 @@
-package roomescape.reservation.controller;
-
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import jakarta.validation.Valid;
-import roomescape.reservation.dto.request.ReservationTimeRequest;
-import roomescape.reservation.dto.response.ReservationTimeResponse;
-import roomescape.reservation.service.ReservationTimeService;
-
-@RequestMapping("/times")
-@RestController
-public class ReservationTimeController {
-
-    private final ReservationTimeService reservationTimeService;
-
-    public ReservationTimeController(final ReservationTimeService reservationTimeService) {
-        this.reservationTimeService = reservationTimeService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationTimeResponse>> readAllReservationTimes() {
-        List<ReservationTimeResponse> responses = reservationTimeService.getAll();
-
-        return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponse> create(@Valid @RequestBody final ReservationTimeRequest request) {
-        ReservationTimeResponse response = reservationTimeService.create(request);
-
-        return ResponseEntity.created(URI.create("/times/" + response.id()))
-                .body(response);
-    }
-
-    @DeleteMapping("/{timeId}")
-    public ResponseEntity<Void> delete(@PathVariable("timeId") final Long timeId) {
-        reservationTimeService.delete(timeId);
-
-        return ResponseEntity.noContent().build();
-    }
-
-}
