--- src/main/java/roomescape/reservation/controller/ThemeController.java
@@ -1,59 +1,0 @@
-package roomescape.reservation.controller;
-
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import jakarta.validation.Valid;
-import roomescape.reservation.dto.request.ThemeRequest;
-import roomescape.reservation.dto.response.ThemeResponse;
-import roomescape.reservation.service.ThemeService;
-
-@RequestMapping("/themes")
-@RestController
-public class ThemeController {
-
-    private final ThemeService themeService;
-
-    public ThemeController(final ThemeService themeService) {
-        this.themeService = themeService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ThemeResponse>> readAllThemes() {
-        List<ThemeResponse> responses = themeService.getAll();
-
-        return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("/popular")
-    public ResponseEntity<List<ThemeResponse>> readPopularThemes() {
-        List<ThemeResponse> responses = themeService.getPopularThemes();
-
-        return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping
-    public ResponseEntity<ThemeResponse> create(@Valid @RequestBody final ThemeRequest request) {
-        ThemeResponse response = themeService.create(request);
-
-        return ResponseEntity.created(URI.create("/themes/" + response.id()))
-                .body(response);
-    }
-
-    @DeleteMapping("/{themeId}")
-    public ResponseEntity<Void> delete(@PathVariable("themeId") final Long themeId) {
-        themeService.delete(themeId);
-
-        return ResponseEntity.noContent().build();
-    }
-
-}
