--- src/main/java/roomescape/reservation/dto/request/ThemeRequest.java
@@ -1,16 +1,0 @@
-package roomescape.reservation.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import roomescape.reservation.domain.Theme;
-
-public record ThemeRequest(
-        @NotBlank String name,
-        @NotBlank String description,
-        @Size(max = 255) @NotBlank String thumbnail
-) {
-
-    public Theme toEntity() {
-        return new Theme(name, description, thumbnail);
-    }
-}
