--- src/main/java/roomescape/reservation/dto/response/ThemeResponse.java
@@ -1,15 +1,0 @@
-package roomescape.reservation.dto.response;
-
-import roomescape.reservation.domain.Theme;
-
-public record ThemeResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnail
-) {
-
-    public static ThemeResponse from(final Theme theme) {
-        return new ThemeResponse(theme.getId(), theme.getName(), theme.getDescription(), theme.getThumbnail());
-    }
-}
