--- src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -1,62 +1,0 @@
-package roomescape.reservation.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import roomescape.common.exception.AlreadyInUseException;
-import roomescape.common.exception.EntityNotFoundException;
-import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.domain.ReservationTimeId;
-import roomescape.reservation.dto.request.ReservationTimeRequest;
-import roomescape.reservation.dto.response.ReservationTimeResponse;
-import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.repository.ReservationTimeRepository;
-
-@Service
-public class ReservationTimeService {
-
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ReservationRepository reservationRepository;
-
-    public ReservationTimeService(
-            final ReservationTimeRepository reservationTimeRepository,
-            final ReservationRepository reservationRepository
-    ) {
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.reservationRepository = reservationRepository;
-    }
-
-    public List<ReservationTimeResponse> getAll() {
-        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
-
-        return reservationTimes.stream()
-                .map(ReservationTimeResponse::from)
-                .toList();
-    }
-
-    @Transactional
-    public ReservationTimeResponse create(final ReservationTimeRequest request) {
-        if (reservationTimeRepository.existsByStartAt(request.startAt())) {
-            throw new AlreadyInUseException("Reservation time already exists");
-        }
-        ReservationTime reservationTime = request.toEntity();
-
-        ReservationTime savedReservationTime = reservationTimeRepository.save(reservationTime);
-
-        return ReservationTimeResponse.from(savedReservationTime);
-    }
-
-    @Transactional
-    public void delete(final Long id) {
-        ReservationTimeId timeId = new ReservationTimeId(id);
-        if (reservationRepository.existsByTimeId(timeId)) {
-            throw new AlreadyInUseException("Reservation is already in use");
-        }
-        if (!reservationTimeRepository.existsById(timeId)) {
-            throw new EntityNotFoundException("존재하지 않는 예약 시간입니다.");
-        }
-        reservationTimeRepository.deleteById(timeId);
-    }
-}
